# Use separate mypy caches for TestTypeHints cases