@samestep 2021-01-14
> Do we also need to update ".gitignore" to reflect the mypy cache directory changes?
> Also, maybe we should cache everything under ".mypy_cache", and put doc cache in ".mypy_cache/doc", example cache in ".mypy_cache/examples", and strict cache in ".mypy_cache/strict"?

@xuzhao9 Good point, I completely forgot about that! At least for the `mypy` version I'm running (0.782), it looks like each of the cache directories automatically includes a `.gitignore` file with these contents:

```gitignore
# Automatically created by mypy
*
```

So I see a few options:

1. leave this PR it as-is
2. remove `.mypy_cache` from our `.gitignore` since it's unnecessary
3. change all the `--cache-dir`s to be subdirectories of `.mypy_cache` as you suggested

Any preference?