@walterddr 2021-01-14
change all the --cache-dirs to be subdirectories of .mypy_cache as you suggested sounds good to me
also circle seems to have some issue today so please also make sure it can work with the CI version of mypy (0.770)