@samestep 2021-01-14
Hmm interesting... I just downgraded to `mypy` 0.770, cleared my cache dir(s), and ran again, and no `.gitignore` files were generated, even though [0.770 was supposed to be the `mypy` version that added that feature](https://mypy-lang.blogspot.com/2020/03/mypy-0770-released.html). In any case, this PR still leaves `.mypy_cache` in our own `.gitignore`, so we're fine by putting everything in subdirectories under that.